server/covid_app/requirements1.txt
asgiref==3.4.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.8
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.0.3
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
Django==3.2.9
djangorestframework==3.12.4
Flask==2.0.2
fonttools==4.28.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
MarkupSafe==2.0.1
mathparse==0.1.2
matplotlib==3.5.0
murmurhash==1.0.6
nltk==3.6.5
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pathy==0.6.1
Pillow==8.4.0
Pint==0.18
preshed==3.0.6
pydantic==1.8.2
pymongo==3.12.1
pyparsing==3.0.6
python-dateutil==2.7.5
pytz==2021.3
PyYAML==3.13
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.2.19
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.0.0
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2